Update .cicd submodule to include check-release-fileset action
Updated .cicd submodule from commit 12db44c to 87f7875 to include the
new check-release-fileset action added in wamp-cicd.

This is required for the release workflow to use the new action:
  uses: wamp-proto/wamp-cicd/actions/check-release-fileset@main

Without this submodule update, the workflow would fail with:
  "Error: Unable to resolve action wamp-proto/wamp-cicd/actions/check-release-fileset@main"

Submodule change: 12db44c..87f7875
- 87f7875 Add check-release-fileset action for PyPI upload validation

This completes the v25.10.2 release preparation.